Creating a search handler
